+++ .eslintrc.js
@@ -1,17 +1,16 @@
 module.exports = {
-	env: {
-		es6: true,
-		node: true
-	},
-	extends: "eslint:recommended",
-	parserOptions: {
-		ecmaVersion: 2018
-	},
-	rules: {
-		"no-unused-vars": ["warn"],
-		indent: ["error", "tab", { SwitchCase: 1 }],
-		"linebreak-style": ["error", "unix"],
-		quotes: ["error", "double"],
-		semi: ["error", "always"]
-	}
+    env: {
+        es6: true,
+        node: true,
+    },
+    extends: "eslint:recommended",
+    parserOptions: {
+        ecmaVersion: 2018,
+    },
+    rules: {
+        "no-unused-vars": ["warn"],
+        "linebreak-style": ["error", "unix"],
+        quotes: ["error", "double"],
+        semi: ["error", "always"],
+    },
 };