+++ index.js
@@ -1,5 +1,6 @@
 const postcss = require("postcss");
 const meta = require("./package.json");
+const mesh = postcss.root();
 const name = meta.name;
 const settings = require("./lib/settings.json");
 const version = meta.version;
@@ -9,17 +10,9 @@ const getDisplaySettings = require("./utils/getDisplaySettings");
 const getCalcedContainerWidth = require("./utils/getCalcedContainerWidth");
 const getInlineSettings = require("./utils/getInlineSettings");
 const valueConversion = require("./lib/valueConversion.json");
-const basicComponents = [
-	"push",
-	"push-basic",
-	"pull",
-	"pull-basic",
-	"column",
-	"offset",
-	"void:after"
-];
+const basicComponents = ["push", "push-basic", "pull", "pull-basic", "column", "offset", "void:after"];
 
-function updateSettings(obj) {
+function updateSettings(obj, viewportRelevant) {
 	// columnAlign
 	if ("column-align" in obj) settings.columnAlign = obj["column-align"];
 
@@ -27,25 +20,18 @@ function updateSettings(obj) {
 	settings.displaySettings = getDisplaySettings(obj);
 
 	// void
-	settings.void.display.value =
-		settings.displayType.value === "flex" ? "flex" : "block";
-	settings.void["font-size"].value =
-		settings.displayType.value === "inline-block" ? "0" : undefined;
+	settings.void.display.value = settings.displayType.value === "flex" ? "flex" : "block";
+	settings.void["font-size"].value = settings.displayType.value === "inline-block" ? "0" : undefined;
 
 	// gutter
-	if ("gutter" in obj)
-		settings.gutter =
-			parseInt(obj.gutter.substring(0, obj.gutter.length - 1)) / 2;
+	if ("gutter" in obj) settings.gutter = parseInt(obj.gutter.substring(0, obj.gutter.length - 1)) / 2;
 
 	// gutterOnOutside
-	if ("gutter-on-outside" in obj)
-		settings.gutterOnOutside =
-			obj["gutter-on-outside"] === "true" ? true : false;
+	if ("gutter-on-outside" in obj) settings.gutterOnOutside = obj["gutter-on-outside"] === "true" ? true : false;
 
 	// responsiveGutter
 	if ("responsive-gutter" in obj) {
-		settings.responsiveGutter =
-			obj["responsive-gutter"] === "true" ? true : false;
+		settings.responsiveGutter = obj["responsive-gutter"] === "true" ? true : false;
 
 		settings.gutterUnit = settings.responsiveGutter ? "%" : "px";
 	}
@@ -56,21 +42,42 @@ function updateSettings(obj) {
 		settings.calcedContainerWidth = getCalcedContainerWidth(obj);
 	}
 
+	// containerBaseWidthUnit
+	if ("container-base-width-unit" in obj && !viewportRelevant) settings.containerBaseWidthUnit = obj["container-base-width-unit"];
+
 	// viewportWidth
-	if ("viewport" in obj)
-		settings.viewportWidth = parseInt(
-			obj.viewport.substring(0, obj.viewport.length - 1)
-		);
+	if ("viewport" in obj) settings.viewportWidth = parseInt(obj.viewport.substring(0, obj.viewport.length - 1));
 
 	// viewportName
 	settings.viewportName = obj.name;
 
 	// columnCount
-	if ("column-count" in obj)
-		settings.columnCount = parseInt(obj["column-count"]);
+	if ("column-count" in obj) settings.columnCount = parseInt(obj["column-count"]);
 
 	// columnSingleWidth
 	settings.columnSingleWidth = 100 / settings.columnCount;
+
+	// naming
+	const namingProps = ["column", "offset", "void", "container", "push", "pull"];
+	for (let i = 0; i < namingProps.length; i++) {
+		const namingProp = `naming-${namingProps[i]}`;
+		if (namingProp in obj && !viewportRelevant) settings[namingProp] = obj[namingProp];
+	}
+
+	// use-name-prefix
+	if ("use-name-prefix" in obj && !viewportRelevant) settings["use-name-prefix"] = obj["use-name-prefix"] == "true";
+
+	// excludes
+	const excludes = ["columns", "offsets", "pulls", "pushes"];
+	for (let i = 0; i < excludes.length; i++) {
+		const exclude = `exclude-${excludes[i]}`;
+		if (exclude in obj) settings[exclude] = obj[exclude].split(",").map(col => parseInt(col));
+	}
+
+	// debug-style
+	if ("debug" in obj && !viewportRelevant) settings.debug.enabled = obj["debug"] == "true";
+	if ("debug-property" in obj && !viewportRelevant) settings.debug.style.prop = obj["debug-property"];
+	if ("debug-value" in obj && !viewportRelevant) settings.debug.style.value = obj["debug-value"];
 }
 
 function updateColumnWidth(fac) {
@@ -79,18 +86,14 @@ function updateColumnWidth(fac) {
 
 function getGutterValue(property, referenceWidth) {
 	let value = settings[property.options.globalKey];
-	value = settings.responsiveGutter
-		? (value / Math.floor(referenceWidth)) * 100
-		: value;
+	value = settings.responsiveGutter ? (value / Math.floor(referenceWidth)) * 100 : value;
 	value = `${value}${settings.gutterUnit}`;
 	return value;
 }
 
 function getAtRule() {
 	const atRule = postcss.atRule();
-	atRule.name = `media (${settings.queryCondition.value} : ${
-		settings.viewportWidth
-	}px)`;
+	atRule.name = `media (${settings.queryCondition.value} : ${settings.viewportWidth}px)`;
 
 	return atRule;
 }
@@ -99,9 +102,7 @@ function getPropValue(component, property) {
 	let value;
 
 	function defaultValue() {
-		value = property.options["value"]
-			? property.options.value
-			: settings[property.options.globalKey];
+		value = property.options["value"] ? property.options.value : settings[property.options.globalKey];
 	}
 
 	if (basicComponents.indexOf(component) >= 0) {
@@ -111,6 +112,9 @@ function getPropValue(component, property) {
 			case "container":
 				if (property.name.indexOf("padding") >= 0) {
 					value = settings.gutterOnOutside ? `${settings.gutter}px` : undefined;
+				} else if (property.name === "width") {
+					defaultValue();
+					value += settings.containerBaseWidthUnit;
 				} else {
 					defaultValue();
 				}
@@ -120,36 +124,21 @@ function getPropValue(component, property) {
 					let percentage = property.index / settings.columnCount;
 					let fac = 1 / percentage;
 					value = settings.gutterOnOutside
-						? getGutterValue(
-								property,
-								settings.calcedContainerWidth / fac - settings.gutter * 2
-						  )
-						: getGutterValue(
-								property,
-								(settings.calcedContainerWidth + settings.gutter * 2) / fac -
-									settings.gutter * 2
-						  );
+						? getGutterValue(property, settings.calcedContainerWidth / fac - settings.gutter * 2)
+						: getGutterValue(property, (settings.calcedContainerWidth + settings.gutter * 2) / fac - settings.gutter * 2);
 					value = value.substring(0, value.length - 1);
 					value = `0 -${value}${settings.gutterUnit}`;
-					value = settings.responsiveGutter
-						? value
-						: `0 -${settings.gutter}${settings.gutterUnit}`;
+					value = settings.responsiveGutter ? value : `0 -${settings.gutter}${settings.gutterUnit}`;
 				}
 				break;
 			case "void":
 				{
 					if (property.name.indexOf("margin") >= 0) {
 						value = settings.gutterOnOutside
-							? getGutterValue(
-									property,
-									settings.calcedContainerWidth - settings.gutter * 2
-							  )
+							? getGutterValue(property, settings.calcedContainerWidth - settings.gutter * 2)
 							: getGutterValue(property, settings.calcedContainerWidth);
 						value = `-${value}`;
-					} else if (
-						(property.name == "flex-wrap" || property.name == "align-items") &&
-						settings.displayType.value !== "flex"
-					) {
+					} else if ((property.name == "flex-wrap" || property.name == "align-items") && settings.displayType.value !== "flex") {
 						value = undefined;
 					} else if (property.name.indexOf("align-items") >= 0) {
 						value = valueConversion[settings[property.options.globalKey]];
@@ -161,10 +150,7 @@ function getPropValue(component, property) {
 			case "column:padding":
 				value = settings.gutterOnOutside
 					? getGutterValue(property, settings.calcedContainerWidth)
-					: getGutterValue(
-							property,
-							settings.calcedContainerWidth + settings.gutter * 2
-					  );
+					: getGutterValue(property, settings.calcedContainerWidth + settings.gutter * 2);
 				break;
 			case "column:nested":
 				{
@@ -172,25 +158,15 @@ function getPropValue(component, property) {
 					let fac = 1 / percentage;
 					value = settings.gutterOnOutside
 						? getGutterValue(property, settings.calcedContainerWidth / fac)
-						: getGutterValue(
-								property,
-								(settings.calcedContainerWidth + settings.gutter * 2) / fac
-						  );
+						: getGutterValue(property, (settings.calcedContainerWidth + settings.gutter * 2) / fac);
 					value = value.substring(0, value.length - 1);
 					value = `0 ${value}${settings.gutterUnit}`;
-					value = settings.responsiveGutter
-						? value
-						: `0 ${settings.gutter}${settings.gutterUnit}`;
+					value = settings.responsiveGutter ? value : `0 ${settings.gutter}${settings.gutterUnit}`;
 				}
 				break;
 			case "column-basic":
 				{
-					const excludeFlexProps = [
-						"vertical-align",
-						"display",
-						"min-height",
-						"float"
-					];
+					const excludeFlexProps = ["vertical-align", "display", "min-height", "float"];
 					const excludeInlineBlockProps = ["vertical-align", "display"];
 					const excludeFloatProps = ["min-height", "float"];
 					let excludeProps;
@@ -210,10 +186,7 @@ function getPropValue(component, property) {
 					if (property.name.indexOf("padding") >= 0) {
 						value = settings.gutterOnOutside
 							? getGutterValue(property, settings.calcedContainerWidth)
-							: getGutterValue(
-									property,
-									settings.calcedContainerWidth + settings.gutter * 2
-							  );
+							: getGutterValue(property, settings.calcedContainerWidth + settings.gutter * 2);
 					} else {
 						defaultValue();
 					}
@@ -229,12 +202,14 @@ function getComponentRules(viewport, options) {
 	const component = options.component;
 	const props = settings[component];
 	const rule = postcss.rule();
+	const debugRule = postcss.rule();
+	debugRule.selector = options.selector;
 	rule.selector = options.selector;
 
 	for (const key in props) {
 		const property = {
 			name: key,
-			options: props[key]
+			options: props[key],
 		};
 		if (options.index) property.index = options.index;
 		const propOptions = props[property.name];
@@ -249,43 +224,118 @@ function getComponentRules(viewport, options) {
 			rule.append(
 				postcss.decl({
 					prop: property.name,
-					value: value
+					value: value,
 				})
 			);
+		}
+
+		if (options.drawDebug) {
+			debugRule.append(postcss.decl(settings.debug.style));
+			mesh.append(debugRule);
 		}
 	}
 	return rule;
 }
 
+function getSelectorByType(type, data = {}) {
+	const namingPatterns = {
+		"naming-column": "|NAME||COLUMN|",
+		"naming-column-span": "|NAME||COLUMN|-|SPAN|",
+		"naming-column-mq": "|NAME||COLUMN|-|MQ|",
+		"naming-column-mq-span": "|NAME||COLUMN|-|MQ|-|SPAN|",
+		"naming-offset": "|NAME||OFFSET|",
+		"naming-offset-span": "|NAME||OFFSET|-|SPAN|",
+		"naming-offset-mq": "|NAME||OFFSET|-|MQ|",
+		"naming-offset-mq-span": "|NAME||OFFSET|-|MQ|-|SPAN|",
+		"naming-container": "|NAME||CONTAINER|",
+		"naming-void": "|NAME||VOID|",
+		"naming-push": "|NAME||PUSH|",
+		"naming-push-span": "|NAME||PUSH|-|SPAN|",
+		"naming-push-mq-span": "|NAME||PUSH|-|MQ|-|SPAN|",
+		"naming-pull": "|NAME||PULL|",
+		"naming-pull-span": "|NAME||PULL|-|SPAN|",
+		"naming-pull-mq-span": "|NAME||PULL|-|MQ|-|SPAN|",
+	};
+
+	const useNamePrefix = settings["use-name-prefix"];
+
+	const haveMQ = data.mq != null;
+	const haveSpan = data.span != null;
+	const haveSpanMQ = haveSpan && haveMQ;
+
+	let pattern;
+	switch (true) {
+		case haveSpanMQ: {
+			pattern = namingPatterns[`naming-${type}-mq-span`];
+			break;
+		}
+		case haveMQ: {
+			pattern = namingPatterns[`naming-${type}-mq`];
+			break;
+		}
+		case haveSpan: {
+			pattern = namingPatterns[`naming-${type}-span`];
+			break;
+		}
+		default: {
+			pattern = namingPatterns[`naming-${type}`];
+		}
+	}
+
+	let selector = pattern.replace(/\|SPAN\|/gm, data.span).replace(/\|MQ\|/gm, data.mq);
+
+	selector = selector
+		.replace(/\|COLUMN\|/gm, settings["naming-column"])
+		.replace(/\|OFFSET\|/gm, settings["naming-offset"])
+		.replace(/\|VOID\|/gm, settings["naming-void"])
+		.replace(/\|CONTAINER\|/gm, settings["naming-container"])
+		.replace(/\|PUSH\|/gm, settings["naming-push"])
+		.replace(/\|PULL\|/gm, settings["naming-pull"]);
+
+	if (useNamePrefix) {
+		selector = selector.replace(/\|NAME\|/gm, `${settings.name}-`);
+	} else {
+		selector = selector.replace(/\|NAME\|/gm, "");
+	}
+
+	return selector;
+}
+
+function excludeSpanByType(type, span) {
+	return settings[`exclude-${type}`].includes(span);
+}
+
 function getRules(grid) {
-	updateSettings(grid);
+	updateSettings(grid, false);
 	const rules = [];
+
+	const debug = settings.debug.enabled && process.env.NODE_ENV !== "production";
 
 	rules.push(
 		// container
 		getComponentRules(grid, {
 			component: "container",
-			selector: `.${settings.name}-container`
+			selector: `.${getSelectorByType("container")}`,
 		}),
 		// void
 		getComponentRules(grid, {
 			component: "void",
-			selector: `.${settings.name}-void`
+			selector: `.${getSelectorByType("void")}`,
 		}),
 		// push-basic
 		getComponentRules(grid, {
 			component: "push-basic",
-			selector: `[class*="${settings.name}-push"]`
+			selector: `[class*="${getSelectorByType("push")}"]`,
 		}),
 		// pull-basic
 		getComponentRules(grid, {
 			component: "pull-basic",
-			selector: `[class*="${settings.name}-pull"]`
+			selector: `[class*="${getSelectorByType("pull")}"]`,
 		}),
 		// column-basic
 		getComponentRules(grid, {
 			component: "column-basic",
-			selector: `[class*="${settings.name}-column"]`
+			selector: `[class*="${getSelectorByType("column", { name: settings.name })}"]`,
 		})
 	);
 
@@ -294,128 +344,191 @@ function getRules(grid) {
 		rules.push(
 			getComponentRules(grid, {
 				component: "void:after",
-				selector: `.${settings.name}-void:after`
+				selector: `.${getSelectorByType("void")}:after`,
 			})
 		);
 	}
+
+	let excludeIt = false;
+	let drawDebug = false;
 
 	for (let i = 0; i <= settings.columnCount; i++) {
 		updateColumnWidth(i);
 
-		if (i !== 0) {
+		excludeIt = excludeSpanByType("columns", i);
+		drawDebug = excludeIt && debug;
+
+		if (i !== 0 && (!excludeIt || debug)) {
 			rules.push(
 				// column
 				getComponentRules(grid, {
 					component: "column",
-					selector: `.${settings.name}-column-${i}`
+					selector: `.${getSelectorByType("column", { span: i })}`,
+					drawDebug,
 				}),
 				// column-x column
 				getComponentRules(grid, {
 					component: "column:nested",
-					selector: `[class*="${settings.name}-column-${i}"] [class*="${
-						settings.name
-					}-column"]`,
-					index: i
+					selector: `.${getSelectorByType("column", {
+						span: i,
+					})} [class*="${getSelectorByType("column", { name: settings.name })}"]`,
+					index: i,
+					drawDebug,
 				}),
 				// column-x void
 				getComponentRules(grid, {
 					component: "void:nested",
-					selector: `[class*="${settings.name}-column-${i}"] .${
-						settings.name
-					}-void`,
-					index: i
+					selector: `.${getSelectorByType("column", { span: i })} .${getSelectorByType("void")}`,
+					index: i,
+					drawDebug,
 				})
 			);
 		}
 
-		rules.push(
-			// push
-			getComponentRules(grid, {
-				component: "push",
-				selector: `.${settings.name}-push-${i}`
-			}),
-			// pull
-			getComponentRules(grid, {
-				component: "pull",
-				selector: `.${settings.name}-pull-${i}`
-			}),
-			// offset
-			getComponentRules(grid, {
-				component: "offset",
-				selector: `.${settings.name}-offset-${i}`
-			})
-		);
+		excludeIt = excludeSpanByType("pushes", i);
+		drawDebug = excludeIt && debug;
+
+		if (!excludeIt || debug) {
+			rules.push(
+				// push
+				getComponentRules(grid, {
+					component: "push",
+					selector: `.${getSelectorByType("push", { span: i })}`,
+					drawDebug,
+				})
+			);
+		}
+
+		excludeIt = excludeSpanByType("pulls", i);
+		drawDebug = excludeIt && debug;
+
+		if (!excludeIt || debug) {
+			rules.push(
+				// pull
+				getComponentRules(grid, {
+					component: "pull",
+					selector: `.${getSelectorByType("pull", { span: i })}`,
+					drawDebug,
+				})
+			);
+		}
+
+		excludeIt = excludeSpanByType("offsets", i);
+		drawDebug = excludeIt && debug;
+
+		if (!excludeIt || debug) {
+			rules.push(
+				// offset
+				getComponentRules(grid, {
+					component: "offset",
+					selector: `.${getSelectorByType("offset", { span: i })}`,
+					drawDebug,
+				})
+			);
+		}
 	}
 
 	for (const key in grid.sortedViewports) {
 		const curViewport = grid.sortedViewports[key];
-		updateSettings(curViewport);
+		updateSettings(curViewport, true);
 		const atRule = getAtRule();
 
 		atRule.append(
 			// container
 			getComponentRules(curViewport, {
 				component: "container",
-				selector: `.${settings.name}-container`
+				selector: `.${getSelectorByType("container")}`,
 			}),
 			// void
 			getComponentRules(curViewport, {
 				component: "void",
-				selector: `.${settings.name}-void`
+				selector: `.${getSelectorByType("void")}`,
 			}),
 			// column-basic
 			getComponentRules(curViewport, {
 				component: "column-basic",
-				selector: `[class*="${settings.name}-column-${settings.viewportName}"]`
+				selector: `[class*="${getSelectorByType("column", { mq: settings.viewportName })}"]`,
 			})
 		);
 
 		for (let i = 0; i <= settings.columnCount; i++) {
 			updateColumnWidth(i);
 
-			if (i !== 0) {
+			excludeIt = excludeSpanByType("columns", i);
+			drawDebug = excludeIt && debug;
+
+			if (i !== 0 && (!excludeIt || debug)) {
 				atRule.append(
 					// column
 					getComponentRules(curViewport, {
 						component: "column",
-						selector: `.${settings.name}-column-${settings.viewportName}-${i}`
+						selector: `.${getSelectorByType("column", { span: i, mq: settings.viewportName })}`,
+						drawDebug,
 					}),
 					// column-x column
 					getComponentRules(grid, {
 						component: "column:nested",
-						selector: `[class*="${settings.name}-column-${
-							settings.viewportName
-						}-${i}"] [class*="${settings.name}-column"]`,
-						index: i
+						selector: `.${getSelectorByType("column", {
+							span: i,
+							mq: settings.viewportName,
+						})} [class*="${getSelectorByType("column")}"]`,
+						index: i,
+						drawDebug,
 					}),
 					// column-x void
 					getComponentRules(grid, {
 						component: "void:nested",
-						selector: `[class*="${settings.name}-column-${
-							settings.viewportName
-						}-${i}"] .${settings.name}-void`,
-						index: i
+						selector: `.${getSelectorByType("column", {
+							span: i,
+							mq: settings.viewportName,
+						})} .${getSelectorByType("void")}`,
+						index: i,
+						drawDebug,
 					})
 				);
 			}
 
-			atRule.append(
-				// push
-				getComponentRules(curViewport, {
-					component: "push",
-					selector: `.${settings.name}-push-${settings.viewportName}-${i}`
-				}),
-				// pull
-				getComponentRules(curViewport, {
-					component: "pull",
-					selector: `.${settings.name}-pull-${settings.viewportName}-${i}`
-				}),
-				// offset
-				getComponentRules(curViewport, {
-					component: "offset",
-					selector: `.${settings.name}-offset-${settings.viewportName}-${i}`
-				})
-			);
+			excludeIt = excludeSpanByType("pushes", i);
+			drawDebug = excludeIt && debug;
+
+			if (!excludeIt || debug) {
+				atRule.append(
+					// push
+					getComponentRules(curViewport, {
+						component: "push",
+						selector: `.${getSelectorByType("push", { mq: settings.viewportName, span: i })}`,
+						drawDebug,
+					})
+				);
+			}
+
+			excludeIt = excludeSpanByType("pulls", i);
+			drawDebug = excludeIt && debug;
+
+			if (!excludeIt || debug) {
+				atRule.append(
+					// pull
+					getComponentRules(curViewport, {
+						component: "pull",
+						selector: `.${getSelectorByType("pull", { mq: settings.viewportName, span: i })}`,
+						drawDebug,
+					})
+				);
+			}
+
+			excludeIt = excludeSpanByType("offsets", i);
+			drawDebug = excludeIt && debug;
+
+			if (!excludeIt || debug) {
+				atRule.append(
+					// offset
+					getComponentRules(curViewport, {
+						component: "offset",
+						selector: `.${getSelectorByType("offset", { span: i, mq: settings.viewportName })}`,
+						drawDebug,
+					})
+				);
+			}
 		}
 
 		rules.push(atRule);
@@ -428,9 +541,6 @@ module.exports = postcss.plugin("postcss-mesh", function() {
 	return function(input) {
 		// inline css settings
 		let inlineSettings = {};
-
-		// generated CSS
-		const mesh = postcss.root();
 
 		// generate styles for base classes
 		function generateCSS() {
@@ -447,21 +557,17 @@ module.exports = postcss.plugin("postcss-mesh", function() {
 				if (curGrid.name) settings.name = curGrid.name;
 
 				// set queryCondition
-				if (curGrid["query-condition"])
-					settings.queryCondition.value = curGrid["query-condition"];
+				if (curGrid["query-condition"]) settings.queryCondition.value = curGrid["query-condition"];
 
 				// set displayType
-				if (
-					"display-type" in curGrid &&
-					settings.displayType.options.indexOf(curGrid["display-type"]) > -1
-				) {
+				if ("display-type" in curGrid && settings.displayType.options.indexOf(curGrid["display-type"]) > -1) {
 					settings.displayType.value = curGrid["display-type"];
 				}
 
 				if (JSON.parse(curGrid["compile"])) mesh.append(getRules(curGrid));
 			}
 
-			input.prepend(mesh);
+			input.append(mesh);
 		}
 
 		// main init